chore: update docs. update dependencies. add make file.
